Make spec files and executables explicit lists, rename files, cl command working
