+++ lib/tp_cli.rb
@@ -1,12 +1,12 @@
 require 'typhoeus'
-require "tp-cli/version"
+require "tp_cli/version"
 
 module TpCommandLine
   class ActivityTrack
-
+    #determine user activity and user feedback or what to POST
     def det_activity
       if ARGV.empty?
-        puts "\n Please specificy action: ./tp_cli.rb [note] or ./tp_cli.rb [cwd]"
+        puts "\n Please specificy action: bin/tp_cli [note] or bin/tp_cli [cwd]"
 
       elsif ARGV[0] == "note" && ARGV[1].is_a?(String)
         post_to_TP(ARGV[1])
@@ -15,7 +15,7 @@ module TpCommandLine
         post_to_TP("Changed working directory")
       end
     end
-
+    #POST to create method in TP Activities controller; create an Activity in db
     def post_to_TP(description)
       request = Typhoeus::Request.new(
         "http://localhost:3000/activities",
